+++ lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:bytebank/transferencia/listagem.dart';
+import 'package:bytebank/screens/listagem.dart';
 import 'package:flutter/material.dart';
 
 void main() {
@@ -6,5 +6,15 @@ void main() {
 }
 
 MaterialApp ByteBankApp() {
-  return MaterialApp(home: ListagemTransferencia());
+  return MaterialApp(
+    home: ListagemTransferencia(),
+    theme: ThemeData(
+      colorScheme: ColorScheme.fromSwatch()
+          .copyWith(primary: Colors.green[900], secondary: Colors.blue[700]),
+      buttonTheme: ButtonThemeData(
+          colorScheme:
+              ColorScheme.fromSwatch().copyWith(secondary: Colors.blue[700]),
+          textTheme: ButtonTextTheme.primary),
+    ),
+  );
 }
